backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

# database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
alembic==1.13.1

# validation
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.9

# background tasks
celery==5.3.6
redis==5.0.1

# image processing
opencv-python==4.9.0.80
pillow==10.2.0

# AI / ML
torch
torchvision
ultralytics==8.1.0

# data processing
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2

# augmentation
albumentations==1.4.3

# visualization
matplotlib==3.8.3
seaborn==0.13.2

# testing
pytest==8.0.2
httpx==0.27.0

# auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# utilities
requests==2.31.0
loguru==0.7.2